EarBridge/services/audio-inference-service/requirements.txt
fastapi
uvicorn
python-multipart
redis
openai
python-dotenv
